requirements.txt
aiogram==3.13.1
motor==3.6.0
pymongo==4.9.2
redis==5.1.1
google-generativeai==0.8.3
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
uvloop==0.20.0
orjson==3.10.7
python-dotenv==1.0.1
